+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+* The `GMH_STAGED_FILES` environment variable containing the path to the repo instead of an empty array if no files are staged.
 
 ### Security